% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CV_Pick}
\alias{CV_Pick}
\title{Pick an optimal lambda from a CV solution path}
\usage{
CV_Pick(
  y_vec,
  x_mat,
  cv_solution_path,
  real_logit_vec,
  kn,
  complex_bound,
  cv_1se = FALSE
)
}
\arguments{
\item{y_vec}{response vector, 0 for control, 1 for case.
n = length(y_vec) is the number of observations.}

\item{x_mat}{covariate matrix, consists of two parts.
dim(x_mat) = (n, h + p * kn)
First h columns are for demographical covariates(can include an intercept term)
Rest columns are for p functional covariates, each being represented by a set of basis functions resulting kn covariates.}

\item{cv_solution_path}{A solution path and related cross validation information.
This is the result from \code{Logistic_FAR_CV_opath}, \code{Logistic_FAR_CV_opath_par}
or \code{Logistic_FAR_CV_Path}.}

\item{real_logit_vec}{Not used in this function}

\item{kn}{number of basis functions for each functional covariates.}

\item{complex_bound}{The upper bound for number of active functional covariates to be considered.
If missing, the whole path will be considered.}

\item{cv_1se}{Logical. Whether the 1se strategy be applied.}
}
\description{
This function picks a best estimation from a CV solution path.
}
\details{
Although the solver function will always return with a selected lambda
This function offers more selecting options.
}
\section{1se Strategy}{

largest value of lambda such that error is within 1 standard error of the maximum likelihood based on CV
}

