% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Logistic_FAR_FLiRTI_Path_Further_Improve}
\alias{Logistic_FAR_FLiRTI_Path_Further_Improve}
\title{Post-selection estimation}
\usage{
Logistic_FAR_FLiRTI_Path_Further_Improve(
  x_mat,
  y_vec,
  h,
  k_n,
  p,
  delta_vec_init,
  eta_stack_init,
  mu1_vec_init,
  mu2,
  a = 1,
  lam = 0.1,
  tol = 10^(-5),
  max_iter = 1000
)
}
\arguments{
\item{x_mat}{covariate matrix, consists of two parts.
dim(x_mat) = (n, h + p * kn)
First h columns are for demographical covariates(can include an intercept term)
Rest columns are for p functional covariates, each being represented by a set of basis functions resulting kn covariates.}

\item{y_vec}{response vector, 0 for control, 1 for case.
n = length(y_vec) is the number of observations.}

\item{h, k_n, p}{dimension information for the dataset(\code{x_mat}).}

\item{delta_vec_init, eta_stack_init, mu1_vec_init}{Initial values for the algorithm.
This function uses these initial values to find out the active functional covariates.
And the post-selection estimation begins with these initial values.}

\item{mu2}{quadratic term in the ADMM algorithm}

\item{a}{parameters for the algorithm. The 1st term in the loss function is
\code{1 / a * loglik}. See Algorithm_Details.pdf
for more information.}

\item{lam}{A scalar for the regularize in ridge penalty form in case of model saturation.}

\item{tol, max_iter}{convergence tolerance and max number of iteration of the algorithm.}
}
\description{
This function performs post-selection estimation on a given solution.
}
