% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{Logistic_FAR_Oracle}
\alias{Logistic_FAR_Oracle}
\title{Oracle estimator of the Logistic FAR problem}
\usage{
Logistic_FAR_Oracle(
  y_vec,
  x_mat,
  h,
  k_n,
  p,
  mu_2,
  h_inv,
  eta_inv_stack,
  delta_init,
  eta_stack_init,
  mu_1_init,
  tol,
  max_iter
)
}
\arguments{
\item{y_vec}{response vector, 0 for control, 1 for case.
n = length(y_vec) is the number of observations.}

\item{x_mat}{covariate matrix, consists of two parts.
dim(x_mat) = (n, h + p * kn)
First \code{h} columns are for demographical covariates.
End user can include an intercept term.
Rest columns are for \code{p} functional covariates, each being represented
by a set of basis functions resulting \code{kn} covariates.
Hence there are \code{p * kn} columns for the functional part.}

\item{h, kn, p}{dimension information for the dataset(\code{x_mat}).}

\item{h_inv, eta_inv_stack}{intermediate variables in the algorithm.
If not provided, then it can be computed from the dataset.}

\item{delta_init, eta_stack_init, mu_1_init}{initial values for the algorithm.}

\item{tol, max_iter}{convergence tolerance and max number of iteration of the algorithm.}

\item{mu2}{quadratic term in the ADMM algorithm}
}
\description{
\code{Logistic_FAR_Oracle} computes the oracle estimator of the Logistic FAR
problem.
}
\details{
In high-dimensional statistics, the oracle estimator is achieved by assuming
the real model structure is known, i.e. an oracle knows what are the relevant
covariates in the model. And the covariate effects are computed through unpenalized
regression.

Basically, this functions performs unpenalized Logistic FAR (with log-contrast constrain)
on the given dataset \code{y_vec} and \code{x_mat}. Hence it is end user who
acts like an oracle to decide which variables will be passed to this function.
Sometimes, this procedure is similar to a post-selection estimation.
}
