% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Replace_Fixef_NA}
\alias{Replace_Fixef_NA}
\title{Get and replace \code{NA} with \code{0} in the fixed-effect coefficients from a Mixed-effect model}
\usage{
Replace_Fixef_NA(in_model)
}
\arguments{
\item{in_model}{An input model, expected from \code{lme4::glmer}.}
}
\value{
The fiexed-effect coefficients of the input model. The dropped
columns (if any) are populated also with coefficients all being 0.
}
\description{
Get and replace \code{NA} with \code{0} in the fixed-effect coefficients from a Mixed-effect model
}
